create a navigation bar component
