+++ src/components/ui/root/Root.jsx
@@ -1,15 +1,12 @@
 import styles from './root.module.css';
-
-import SiteNav from '../site-nav/SiteNav';
-
+import Header from '../header/Header';
 import Breadcrumb from '../breadcrumb/Breadcrumb';
-
 import CardComponent from "../cardComponent/CardComponent";
 
 function Root() {
 	return (
 		<div className={styles.wrapper}>
-			<SiteNav />
+			<Header />
 			<Breadcrumb />
 			<CardComponent/>
 		</div>
