+++ src/components/ui/root/root.module.css
@@ -1,4 +1,5 @@
-.wrapper {
+
+/* .wrapper {
 	display: flex;
 	flex-direction: column;
 	width: 90%;
@@ -8,4 +9,4 @@
 	background: #fff;
 	overflow: hidden;
 	border-radius: 20px;
-}
+} */
